Update docs for new API (#10)
* Update README.
* Update Docc landing page, add Assertion docs.
* Rename files to reinforce operator terminology.
* Add article explaining how to write and test custom operators.
