+++ README.md
@@ -2,12 +2,12 @@
 
 ## What is Gauntlet?
 
-Gauntlet is a collection of testing utility methods that aims to make writing great tests easier and with more helpful and descriptive failure states. It is only intended for inclusion in test cases. It has a dependency on `XCTest` but no external dependencies.
+Gauntlet is an extensible, functional API for writing unit test assertions in Swift, built on top of XCTest. It is only intended for inclusion in test cases. It has a dependency on `XCTest` but no external dependencies.
 
 ## Requirements
 
-- XCode 13.2+
-- Swift 5.5+
+- XCode 14.2+
+- Swift 5.7+
 
 ## Installation
 
@@ -17,7 +17,7 @@ The easiest way to install Gauntlet is by adding a dependency via SPM.
         .package(
             name: "Gauntlet",
             url: "https://github.com/krogerco/gauntlet-ios.git",
-            .upToNextMajor(from: Version(1, 0, 0))
+            .upToNextMajor(from: Version(2, 0, 0))
         )
 ```
 
@@ -31,134 +31,48 @@ The easiest way to install Gauntlet is by adding a dependency via SPM.
         )
 ```
 
-## How to Use Gauntlet
+## Using Gauntlet
 
-Gauntlet augments the standard XCTest unit testing methods to make unit tests more expressive and easier to read. Many of the methods in Gauntlet mirror those present in XCTest (e.g. `XCTAssertEqual`) but additionally include a closure that will be called if the assert passes.
-
-For example, a typical unit test might look like:
+Gauntlet assertions are used in place of the `XCTAssert` functions. You start by creating an assertion:
 
 ```swift
-func testSomething() throws {
-    // Given
-    let model = Model()
+// Given
+let model = Model()
 
-    // When
-    model.doSomething()
+// When
+let result: Result<String, Error> = model.loadContent()
 
-    // Then
-    XCTAssertNotNil(model.file)
-    XCTAssertNotNil(model.file?.url)
-    XCTAssertEqual(model.file?.url?.lastPathComponent, "myFile.json")
-    XCTAssertFalse(model.file!.exists())
+Assert(that: result)
+```
+
+However this assertion doesn't do anything yet. To actually validate the value you need to call an operator on it:
+
+```swift
+let model = Model()
+
+// When
+let result: Result<String, Error> = model.loadContent()
+
+Assert(that: result).isSuccess().isEqualTo("expected content")
+```
+
+An operator is a function defined on an `Assertion` which performs validation on a value and then returns a new Assertion. If an operator's validation fails a test failure will be generated and no subsequent operators on that assertion will be evaluated. In the above example the `isSuccess()` operator validates that the result is a success and provides the associated success value in the returned Assertion.
+
+Gauntlet includes a `then` operator which can be used to break an Assertion out into more assertions that are only run when the original Assertion passes.
+
+```swift
+let model = Model()
+
+// When
+let result: Result<Content, Error> = model.loadContent()
+
+Assert(that: result).isSuccess().then { content in
+    Assert(that: content.id).isEqualTo("expected id")
+    ...
 }
 ```
 
-Unit tests involving optionals are commonly written like the above, either relying on force-unwraps or manual unwrapping of optionals to write all of the needed asserts. This leads to the following problems:
-- Force unwraps make the tests fragile. If the conditions are not met, the test will crash and no other tests will be run. You cannot know how many other tests would have passed or failed.
-- Manual unwrapping clutters up the code, and can lead to needing to write inline ternary statements to get the type needed for an assert.
-- You may need to disable swiftlint rules for your unit tests.
-- The expectation tree cannot be easily understood. if the first assert fails, there is no point in testing the following asserts.
-
-A better version of this might look like the following:
-
-```swift
-func testSomething() throws {
-    // Given
-    let model = Model()
-
-    // When
-    model.doSomething()
-
-    // Then
-    XCTAssertNotNil(model.file)
-    if let file = model.file {
-        XCTAssertNotNil(file.url)
-
-        if let url = file.url {
-            XCTAssertEqual(url.lastPathComponent, "myFile.json")
-        }
-
-        XCTAssertFalse(file.exists())
-    }
-}
-```
-
-The fragility of the force unwraps has been eliminated, at the cost of readability of the code. Control flow is now intermingled with asserts.
-
-Using Gauntlet and the addition of assert closures that are only executed on success, the meaning and expectations of the test become significantly more clear.
-
-```swift
-import Gauntlet
-import XCTest
-
-func testSomething() throws {
-    // Given
-    let model = Model()
-
-    // When
-    model.doSomething()
-
-    // Then
-    XCTAssertNotNil(model.file) { file in
-        XCTAssertNotNil(file.url) { url in
-            XCTAssertEqual(url.lastPathComponent, "myFile.json")
-        }
-        XCTAssertFalse(file.exists())
-    }
-}
-```
-
-When `XCTAssertNotNil` succeeds, the closure is called with the unwrapped value, ready to use in the next assert.
-
-This leads to the following benefits:
-- You can tell at a glance what are the expectations of the test.
-- Optionals are safely unwrapped with minimal visual clutter.
-- Follow-on asserts are only executed when appropriate.
-- Code is compact and readable.
-
-This functionality is made available to many of the assert methods you use today, and can be taken advantage of simply by adding a trailing closure.
-
-For example, this code ensures you can test the contents of specific array indices without crashing if the array is underfilled.
-
-```swift
-XCTAssertEqual(myArray.count, 2) {
-    XCTAssertEqual(myArray[0], "Hello")
-    XCTAssertEqual(myArray[1], "World")
-}
-```
-
-Other asserts also help remove boilerplate code when working with types like `Result`.
-
-```swift
-// Given, When
-let result: Result<Any, Error> = ...
-
-// Then
-XCTAssertSuccess(result, is: String.self) { value in
-    XCTAssertFalse(value.isEmpty)
-}
-```
-
-In this case, `XCTAssertSuccess` eliminates the boilerplate code needed to convert the success type to the desired type and the switch statement that would subsequently needed for `.success` and `.failure`.
-
-There are assert additions for a variety of cases, including: true, false, empty collections, equality, optionals, results, types, and more.
-
-> Note: Each function includes optional parameters for `reporter`, `file`, and `line`. These are hooks to test the functions themselves and in 99% of cases should be left to their default values for best results.
-
-## Why are they named like the XCTest provided methods?
-
-We debated this very point quite a bit when writing Gauntlet. By sticking with the existing method prefixes we gain substantial benefits:
-
-- The discoverability of the enhanced methods is greatly increased. As you type the normal `XCT` prefix, auto-complete will list all possible variations that are available, including the new methods.
-- Readability at the call site is greatly increased. They look and read like the XCTest methods a developer is already familiar with.
-
-In the unlikely event of a name collision in the future , the namespace of the module (either `XCTest` or `Gauntlet`) can be prepended at the callsite to disambiguate. If this were to happen, we would obviously strive to resolve this in a subsequent version of Gauntlet.
-
-## Extensions
-
-Gauntlet also includes some extensions that can be useful when writing tests.
-
-A static property `currentQueueLabel` has been added to `DispatchQueue`. This is useful when testing code that takes a queue that it calls a completion on to verify that the completion was called on the queue that was passed to that code.
+Gauntlet provides a number of operators, and is designed to be extensible so that you can build and test your own operators easily. Check the [docs](#documentation) for more details.
 
 ## Documentation
