+++ Source/Gauntlet/Operators/AssertionOperators.swift
@@ -1,5 +1,5 @@
 //
-//  AssertionAssertions.swift
+//  AssertionOperators.swift
 //
 //  MIT License
 //
