+++ Tests/GauntletTests/Operators/ThrowingOperatorsTests.swift
@@ -1,5 +1,5 @@
 //
-//  ThrowingAssertionsTests.swift
+//  ThrowingOperatorsTests.swift
 //
 //  MIT License
 //
